small/feed_to_telegraph/pyproject.toml
[project]
name = "feed-to-telegraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.14",
    "dotenv>=0.9.9",
    "feedparser>=6.0.11",
    "html-telegraph-poster>=0.5.0",
    "lxml[html-clean]>=5.3.1",
    "python-telegram-bot>=22.0",
    "tqdm>=4.67.1",
]
